Poprawki w oczekiwanym hashu

+++ Client/PeerConnection.cpp
@@ -210,8 +210,8 @@ void PeerConnection::readData()
 			else if (message[0] == '6')
 			{
 				size_t index = read(message.substr(1, sizeof(size_t)));
-				size_t begin = read(message.substr(1 + sizeof(size_t), sizeof(size_t)));
-				size_t length = read(message.substr(1 + 2 * sizeof(size_t), sizeof(size_t)));
+				int begin = read(message.substr(1 + sizeof(size_t), sizeof(size_t)));
+				int length = read(message.substr(1 + 2 * sizeof(size_t), sizeof(size_t)));
 				piece(index, begin, (*mFile)[(const unsigned int)index]->getData().mid((int)begin, (int)length).toStdString());
 				
 			}
